Reorganize fake event source
